requirement.txt
streamlit
torch
torchvision
efficientnet_pytorch
matplotlib
pandas
pillow
